# docs/swagger.json
{
	"components": {
...
		"schemas": {
...
			"IRegisterPayload": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"repeatPassword": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"repeatPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserPayload": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "calculator_server",
		"version": "1.0.0",
		"description": "A simple server for a calculator app written in typescript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Abdulraheem A omeiza"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/user/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"email": "john@doe.com",
											"password": "12345",
											"name": "john doe"
										}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "user registered successfully",
											"data": {
												"email": "john@doe.com",
												"name": "john doe",
												"createdAt": "2021-09-15T17:07:15.000Z",
												"updatedAt": "2021-09-15T17:07:15.000Z",
												"id": "6142b4a3b0d7f4f8a7f8b7f6"
											},
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "user registered successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Used for user registratiion.\r\nSupply the user's name, unique email and password",
				"tags": [
					"USER"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRegisterPayload"
							}
						}
					}
				}
			}
		},
		"/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"email": "john@doe.com",
											"password": "12345"
										}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "user registered successfully",
											"data": {
												"email": "john@doe.com",
												"name": "john doe",
												"createdAt": "2021-09-15T17:07:15.000Z",
												"updatedAt": "2021-09-15T17:07:15.000Z",
												"id": "6142b4a3b0d7f4f8a7f8b7f6"
											},
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "user logged in successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USER"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/user/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"email": "john@doe.com"
										}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "user registered successfully",
											"data": {
												"email": "john@doe.com",
												"name": "john doe",
												"createdAt": "2021-09-15T17:07:15.000Z",
												"updatedAt": "2021-09-15T17:07:15.000Z",
												"id": "6142b4a3b0d7f4f8a7f8b7f6"
											},
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "user logged out successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USER"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/user/change-password": {
			"patch": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"email": "john@doe.com",
											"password": "12345",
											"name": "john doe"
										}
									},
									"Example 2": {
										"value": {
											"success": true,
											"message": "user registered successfully",
											"data": {
												"email": "john@doe.com",
												"name": "john doe",
												"createdAt": "2021-09-15T17:07:15.000Z",
												"updatedAt": "2021-09-15T17:07:15.000Z",
												"id": "6142b4a3b0d7f4f8a7f8b7f6"
											},
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "password changed successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USER"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/user/get-all": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "users fetched successfully",
											"data": [
												{
													"email": "john@doe.com",
													"name": "john doe",
													"createdAt": "2021-09-15T17:07:15.000Z",
													"updatedAt": "2021-09-15T17:07:15.000Z",
													"id": "6142b4a3b0d7f4f8a7f8b7f6"
												}
											],
											"token": "token"
										}
									}
								}
							}
						}
					},
					"201": {
						"description": "users fetched successfully"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"USER"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}